rename css file to fix ssr test on mac
